MockScatterChart factoring (Point Chart) and implementation
